requirements-dev.txt
pytest
requests
pytest-playwright
playwright
psycopg2-binary
